Add files via upload
This project is a simple and easy guide on how to Data model. To create DB and insert data into it.
This project highlights the importance of Data modeling as most data scientists gain data in proper shape and do not need to perform these processes into it.
So this process is the prior step, mostly performed by Data Engineers.